HW9: video: rename StreamSerializer and ObjectStreamSerializer; in PathStorage and FileStorage refactor constructor
