+++ src/com/javaops/webapp/storage/serializer/StreamSerializer.java
@@ -1,4 +1,4 @@
-package com.javaops.webapp.storage.strategy;
+package com.javaops.webapp.storage.serializer;
 
 import com.javaops.webapp.model.Resume;
 
@@ -6,7 +6,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-public interface Strategy {
+public interface StreamSerializer {
 
     Resume doRead(InputStream inputStream) throws IOException;
