+++ test/com/javaops/webapp/storage/FileStorageTest.java
@@ -1,10 +1,10 @@
 package com.javaops.webapp.storage;
 
-import com.javaops.webapp.storage.strategy.ObjectStreamStrategy;
+import com.javaops.webapp.storage.serializer.ObjectStreamSerializer;
 
 public class FileStorageTest extends AbstractStorageTest {
 
     public FileStorageTest() {
-        super(new FileStorage(STORAGE_DIR, new ObjectStreamStrategy()));
+        super(new FileStorage(STORAGE_DIR, new ObjectStreamSerializer()));
     }
 }
