+++ test/com/javaops/webapp/storage/PathStorageTest.java
@@ -1,10 +1,10 @@
 package com.javaops.webapp.storage;
 
-import com.javaops.webapp.storage.strategy.ObjectStreamStrategy;
+import com.javaops.webapp.storage.serializer.ObjectStreamSerializer;
 
 public class PathStorageTest extends AbstractStorageTest {
 
     public PathStorageTest() {
-        super(new PathStorage(STORAGE_DIR.getAbsolutePath(), new ObjectStreamStrategy()));
+        super(new PathStorage(STORAGE_DIR.getAbsolutePath(), new ObjectStreamSerializer()));
     }
 }
